Updated app editor spec
